requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
pytest>=6.2.0
pytest-cov>=2.12.0
pydantic>=1.8.0
pyyaml>=5.4.0
